agents/concierge/requirements.txt
# ── Fetch.ai ──────────────────────────────────────────────────────────────────
uagents==0.24.2
uagents-core==0.4.4

# ── Reasoning pipeline ────────────────────────────────────────────────────────
langgraph>=0.2.0,<0.3
langchain>=0.3.0,<0.4
langchain-anthropic>=0.3.0,<0.4
langchain-openai>=0.2.0,<0.3       # text-embedding-3-small

# ── Database ──────────────────────────────────────────────────────────────────
supabase>=2.4.0,<3.0

# ── HTTP / retries ────────────────────────────────────────────────────────────
httpx>=0.27.0
tenacity>=8.3.0

# ── Config ────────────────────────────────────────────────────────────────────
pydantic>=2.7.0,<3.0
pydantic-settings>=2.3.0,<3.0
python-dotenv>=1.0.0
